fixed dirichlet, disabled restirct for now
