+++ hpgmg/finite_volume/operators/transformers/semantic_transformer.py
@@ -1,7 +1,7 @@
 import ast
 from hpgmg.finite_volume.operators.nodes import RangeNode
 from hpgmg.finite_volume.operators.transformers.transformer_util import get_obj
-from hpgmg.finite_volume.operators.transformers.utility_transformers import get_name
+from hpgmg.finite_volume.operators.transformers.utility_transformers import get_name, eval_node
 
 __author__ = 'nzhang-dev'
 
@@ -28,17 +28,7 @@ class SemanticFinder(ast.NodeTransformer):
         func_name = split[-1]
 
         if func_name not in self.registered or obj_name not in self.namespace:
-            print(obj_name, func_name)
+            #print(obj_name, func_name)
             return self.generic_visit(node)
         func_obj = get_obj(self.namespace, iter_name)
         return RangeNode(iteration_variable_name, func_obj(*[eval_node(arg, self.locals, self.globals) for arg in node.iter.args]), node.body)
-
-def eval_node(node, locals, globals):
-    expr = ast.Expression(node)
-    expr = ast.fix_missing_locations(expr)
-    #print(locals, globals)
-    items = eval(
-        compile(expr, "<string>", "eval"),
-        globals, locals
-    )
-    return items
